feat: refactor basic tools and remove deprecated functions
